requirements.txt
elasticsearch==8.10.0
flake8==4.0
psycopg2-binary==2.9.7
pydantic==2.4.2
redis==5.0.1
ipython
orjson
